Move proxy into sub package
